Chapter 9 Ex 19 - простая фабрика на базе класса
