+++ src/main/java/ru/kulichenkom/ekkel/chapter9/game/GamePlay.java
@@ -1,9 +1,18 @@
 package ru.kulichenkom.ekkel.chapter9.game;
 
+import ru.kulichenkom.ekkel.chapter9.game.factories.SecondGameFactory;
+import ru.kulichenkom.ekkel.chapter9.game.interfaces.Game;
+
+import java.util.Scanner;
+
 public class GamePlay {
+
     public static void main(String[] args) {
-        GameFactory maker = new DiceGameMaker();
-        Game game = maker.makeGame();
+        System.out.println("Введите название игры: COIN или DICE ");
+        Scanner in = new Scanner(System.in);
+        String typeOfGame = in.nextLine().toUpperCase();
+        SecondGameFactory secondGameFactory = new SecondGameFactory();
+        Game game = secondGameFactory.makeGame(TypeOfGame.valueOf(typeOfGame));
         game.tossUp();
     }
 }
