+++ src/main/java/ru/kulichenkom/ekkel/chapter9/game/TypeOfGame.java
@@ -1,0 +1,6 @@
+package ru.kulichenkom.ekkel.chapter9.game;
+
+public enum TypeOfGame {
+    DICE,
+    COIN
+}
