+++ src/main/java/ru/kulichenkom/ekkel/chapter9/game/interfaces/Game.java
@@ -1,0 +1,8 @@
+//Ch9Ex19 Создайте на базе паттерна "Фабричный метод" программную среду,
+//моделирующую броски монет и броски кубиков.
+package ru.kulichenkom.ekkel.chapter9.game.interfaces;
+
+public interface Game {
+    void tossUp();
+}
+
