+++ src/main/java/ru/kulichenkom/ekkel/chapter9/game/interfaces/GameFactory.java
@@ -1,0 +1,5 @@
+package ru.kulichenkom.ekkel.chapter9.game.interfaces;
+
+public interface GameFactory {
+    Game makeGame();
+}
